requirements.txt
chromedriver_binary==108.0.5359.22.0
pandas==1.5.1
python-dotenv==0.21.0
selenium==4.6.0
webdriver_manager==3.8.4
setuptools
packaging